Clean up engine files
