+++ .gitignore
@@ -1,3 +1,4 @@
 *.o
 *.dSYM
+.DS_Store
 42RL
