+++ K4Engine/source/Director.cpp
