+++ K4Engine/source/Scene.cpp
