--- include/KeyCode.hpp
@@ -1,44 +1,0 @@
-#ifndef KEY_CODE_HPP
-#define KEY_CODE_HPP
-
-#include <SDL2/SDL.h>
-
-enum KeyCode
-{
-	KEY_SPACE = SDL_SCANCODE_SPACE,
-	KEY_UP = SDL_SCANCODE_UP,
-	KEY_DOWN = SDL_SCANCODE_DOWN,
-	KEY_LEFT = SDL_SCANCODE_LEFT,
-	KEY_RIGHT = SDL_SCANCODE_RIGHT,
-	KEY_A = SDL_SCANCODE_A,
-	KEY_B = SDL_SCANCODE_B,
-	KEY_C = SDL_SCANCODE_C,
-	KEY_D = SDL_SCANCODE_D,
-	KEY_E = SDL_SCANCODE_E,
-	KEY_F = SDL_SCANCODE_F,
-	KEY_G = SDL_SCANCODE_G,
-	KEY_H = SDL_SCANCODE_H,
-	KEY_I = SDL_SCANCODE_I,
-	KEY_J = SDL_SCANCODE_J,
-	KEY_K = SDL_SCANCODE_K,
-	KEY_L = SDL_SCANCODE_L,
-	KEY_M = SDL_SCANCODE_M,
-	KEY_N = SDL_SCANCODE_N,
-	KEY_O = SDL_SCANCODE_O,
-	KEY_P = SDL_SCANCODE_P,
-	KEY_Q = SDL_SCANCODE_Q,
-	KEY_R = SDL_SCANCODE_R,
-	KEY_S = SDL_SCANCODE_S,
-	KEY_T = SDL_SCANCODE_T,
-	KEY_U = SDL_SCANCODE_U,
-	KEY_W = SDL_SCANCODE_W,
-	KEY_X = SDL_SCANCODE_X,
-	KEY_Y = SDL_SCANCODE_Y,
-	KEY_Z = SDL_SCANCODE_Z,
-	KEY_LCTRL = SDL_SCANCODE_LCTRL,
-	KEY_RCTRL = SDL_SCANCODE_RCTRL,
-	KEY_LSHIFT = SDL_SCANCODE_LSHIFT,
-	KEY_RSHIFT = SDL_SCANCODE_RSHIFT
-};
-
-#endif
